+++ openmls/src/group/mls_group/creation.rs
@@ -1,3 +1,4 @@
+use errors::NewGroupError;
 use openmls_traits::{signatures::Signer, storage::StorageProvider as StorageProviderTrait};
 
 use super::{builder::MlsGroupBuilder, *};
@@ -15,8 +16,9 @@ use crate::{
         psk::{store::ResumptionPskStore, PreSharedKeyId},
         EpochSecrets, InitSecret,
     },
-    storage::{self, OpenMlsProvider},
+    storage::OpenMlsProvider,
     treesync::{
+        errors::{DerivePathError, PublicTreeError},
         node::leaf_node::{Capabilities, LeafNodeParameters},
         RatchetTreeIn,
     },
@@ -115,14 +117,13 @@ impl MlsGroup {
         // If we got a ratchet tree extension in the welcome, we enable it for
         // this group. Note that this is not strictly necessary. But there's
         // currently no other mechanism to enable the extension.
-        let (ratchet_tree, enable_ratchet_tree_extension) =
-            match verifiable_group_info.extensions().ratchet_tree() {
-                Some(extension) => (extension.ratchet_tree().clone(), true),
-                None => match ratchet_tree {
-                    Some(ratchet_tree) => (ratchet_tree, false),
-                    None => return Err(ExternalCommitError::MissingRatchetTree),
-                },
-            };
+        let ratchet_tree = match verifiable_group_info.extensions().ratchet_tree() {
+            Some(extension) => extension.ratchet_tree().clone(),
+            None => match ratchet_tree {
+                Some(ratchet_tree) => ratchet_tree,
+                None => return Err(ExternalCommitError::MissingRatchetTree),
+            },
+        };
 
         let (public_group, group_info) = PublicGroup::from_external(
             provider.crypto(),
@@ -250,13 +251,64 @@ impl ProcessedWelcome {
         let (resumption_psk_store, key_package_bundle) =
             keys_for_welcome(mls_group_config, &welcome, provider)?;
 
-        let (ciphersuite, group_secrets, key_schedule, verifiable_group_info) =
-            crate::group::core_group::new_from_welcome::process_welcome(
-                welcome,
-                &key_package_bundle,
-                provider,
+        let ciphersuite = welcome.ciphersuite();
+        let Some(egs) = welcome.find_encrypted_group_secret(
+            key_package_bundle
+                .key_package()
+                .hash_ref(provider.crypto())?,
+        ) else {
+            return Err(WelcomeError::JoinerSecretNotFound);
+        };
+        if ciphersuite != key_package_bundle.key_package().ciphersuite() {
+            let e = WelcomeError::CiphersuiteMismatch;
+            log::debug!("new_from_welcome {:?}", e);
+            return Err(e);
+        }
+        let group_secrets = GroupSecrets::try_from_ciphertext(
+            key_package_bundle.init_private_key(),
+            egs.encrypted_group_secrets(),
+            welcome.encrypted_group_info(),
+            ciphersuite,
+            provider.crypto(),
+        )?;
+        let psk_secret = {
+            let psks = load_psks(
+                provider.storage(),
                 &resumption_psk_store,
+                &group_secrets.psks,
             )?;
+
+            PskSecret::new(provider.crypto(), ciphersuite, psks)?
+        };
+        let key_schedule = KeySchedule::init(
+            ciphersuite,
+            provider.crypto(),
+            &group_secrets.joiner_secret,
+            psk_secret,
+        )?;
+        let (welcome_key, welcome_nonce) = key_schedule
+            .welcome(provider.crypto(), ciphersuite)
+            .map_err(|_| LibraryError::custom("Using the key schedule in the wrong state"))?
+            .derive_welcome_key_nonce(provider.crypto(), ciphersuite)
+            .map_err(LibraryError::unexpected_crypto_error)?;
+        let verifiable_group_info = VerifiableGroupInfo::try_from_ciphertext(
+            &welcome_key,
+            &welcome_nonce,
+            welcome.encrypted_group_info(),
+            &[],
+            provider.crypto(),
+        )?;
+        if let Some(required_capabilities) =
+            verifiable_group_info.extensions().required_capabilities()
+        {
+            // Also check that our key package actually supports the extensions.
+            // Per spec the sender must have checked this. But you never know.
+            key_package_bundle
+                .key_package()
+                .leaf_node()
+                .capabilities()
+                .supports_required_capabilities(required_capabilities)?;
+        }
 
         Ok(Self {
             mls_group_config: mls_group_config.clone(),
@@ -286,24 +338,139 @@ impl ProcessedWelcome {
     /// Consume the `ProcessedWelcome` and combine it witht he ratchet tree into
     /// a `StagedWelcome`.
     pub fn into_staged_welcome<Provider: OpenMlsProvider>(
-        self,
+        mut self,
         provider: &Provider,
         ratchet_tree: Option<RatchetTreeIn>,
     ) -> Result<StagedWelcome, WelcomeError<Provider::StorageError>> {
-        let group = crate::group::core_group::new_from_welcome::build_staged_welcome(
-            self.verifiable_group_info,
+        // Build the ratchet tree and group
+
+        // Set nodes either from the extension or from the `nodes_option`.
+        // If we got a ratchet tree extension in the welcome, we enable it for
+        // this group. Note that this is not strictly necessary. But there's
+        // currently no other mechanism to enable the extension.
+        let ratchet_tree = match self.verifiable_group_info.extensions().ratchet_tree() {
+            Some(extension) => extension.ratchet_tree().clone(),
+            None => match ratchet_tree {
+                Some(ratchet_tree) => ratchet_tree,
+                None => return Err(WelcomeError::MissingRatchetTree),
+            },
+        };
+
+        // Since there is currently only the external pub extension, there is no
+        // group info extension of interest here.
+        let (public_group, _group_info_extensions) = PublicGroup::from_external(
+            provider.crypto(),
+            provider.storage(),
             ratchet_tree,
-            provider,
-            self.key_package_bundle,
-            self.key_schedule,
-            self.ciphersuite,
-            self.resumption_psk_store,
-            self.group_secrets,
+            self.verifiable_group_info.clone(),
+            ProposalStore::new(),
         )?;
+
+        // Find our own leaf in the tree.
+        let own_leaf_index = public_group
+            .members()
+            .find_map(|m| {
+                if m.signature_key
+                    == self
+                        .key_package_bundle
+                        .key_package()
+                        .leaf_node()
+                        .signature_key()
+                        .as_slice()
+                {
+                    Some(m.index)
+                } else {
+                    None
+                }
+            })
+            .ok_or(WelcomeError::PublicTreeError(
+                PublicTreeError::MalformedTree,
+            ))?;
+
+        let (group_epoch_secrets, message_secrets) = {
+            let serialized_group_context = public_group
+                .group_context()
+                .tls_serialize_detached()
+                .map_err(LibraryError::missing_bound_check)?;
+
+            // TODO #751: Implement PSK
+            self.key_schedule
+                .add_context(provider.crypto(), &serialized_group_context)
+                .map_err(|_| LibraryError::custom("Using the key schedule in the wrong state"))?;
+
+            let epoch_secrets = self
+                .key_schedule
+                .epoch_secrets(provider.crypto(), self.ciphersuite)
+                .map_err(|_| LibraryError::custom("Using the key schedule in the wrong state"))?;
+
+            epoch_secrets.split_secrets(
+                serialized_group_context,
+                public_group.tree_size(),
+                own_leaf_index,
+            )
+        };
+
+        let confirmation_tag = message_secrets
+            .confirmation_key()
+            .tag(
+                provider.crypto(),
+                self.ciphersuite,
+                public_group.group_context().confirmed_transcript_hash(),
+            )
+            .map_err(LibraryError::unexpected_crypto_error)?;
+
+        // Verify confirmation tag
+        if &confirmation_tag != public_group.confirmation_tag() {
+            log::error!("Confirmation tag mismatch");
+            log_crypto!(trace, "  Got:      {:x?}", confirmation_tag);
+            log_crypto!(trace, "  Expected: {:x?}", public_group.confirmation_tag());
+            debug_assert!(false, "Confirmation tag mismatch");
+
+            // in some tests we need to be able to proceed despite the tag being wrong,
+            // e.g. to test whether a later validation check is performed correctly.
+            if !crate::skip_validation::is_disabled::confirmation_tag() {
+                return Err(WelcomeError::ConfirmationTagMismatch);
+            }
+        }
+
+        let message_secrets_store = MessageSecretsStore::new_with_secret(0, message_secrets);
+
+        // Extract and store the resumption PSK for the current epoch.
+        let resumption_psk = group_epoch_secrets.resumption_psk();
+        self.resumption_psk_store
+            .add(public_group.group_context().epoch(), resumption_psk.clone());
+
+        let welcome_sender_index = self.verifiable_group_info.signer();
+        let path_keypairs = if let Some(path_secret) = self.group_secrets.path_secret {
+            let (path_keypairs, _commit_secret) = public_group
+                .derive_path_secrets(
+                    provider.crypto(),
+                    self.ciphersuite,
+                    path_secret,
+                    welcome_sender_index,
+                    own_leaf_index,
+                )
+                .map_err(|e| match e {
+                    DerivePathError::LibraryError(e) => e.into(),
+                    DerivePathError::PublicKeyMismatch => {
+                        WelcomeError::PublicTreeError(PublicTreeError::PublicKeyMismatch)
+                    }
+                })?;
+            Some(path_keypairs)
+        } else {
+            None
+        };
 
         let staged_welcome = StagedWelcome {
             mls_group_config: self.mls_group_config,
-            group,
+            public_group,
+            group_epoch_secrets,
+            own_leaf_index,
+            message_secrets_store,
+            resumption_psk_store: self.resumption_psk_store,
+            verifiable_group_info: self.verifiable_group_info,
+            key_package_bundle: self.key_package_bundle,
+            path_keypairs,
         };
 
         Ok(staged_welcome)
@@ -324,37 +491,29 @@ impl StagedWelcome {
         welcome: Welcome,
         ratchet_tree: Option<RatchetTreeIn>,
     ) -> Result<Self, WelcomeError<Provider::StorageError>> {
-        let (resumption_psk_store, key_package_bundle) =
-            keys_for_welcome(mls_group_config, &welcome, provider)?;
+        let processed_welcome =
+            ProcessedWelcome::new_from_welcome(provider, mls_group_config, welcome)?;
 
-        let group = StagedCoreWelcome::new_from_welcome(
-            welcome,
-            ratchet_tree,
-            key_package_bundle,
-            provider,
-            resumption_psk_store,
-        )?;
-
-        let staged_welcome = StagedWelcome {
-            mls_group_config: mls_group_config.clone(),
-            group,
-        };
-
-        Ok(staged_welcome)
+        processed_welcome.into_staged_welcome(provider, ratchet_tree)
     }
 
     /// Returns the [`LeafNodeIndex`] of the group member that authored the [`Welcome`] message.
     ///
     /// [`Welcome`]: crate::messages::Welcome
     pub fn welcome_sender_index(&self) -> LeafNodeIndex {
-        self.group.welcome_sender_index()
+        self.verifiable_group_info.signer()
     }
 
     /// Returns the [`LeafNode`] of the group member that authored the [`Welcome`] message.
     ///
     /// [`Welcome`]: crate::messages::Welcome
     pub fn welcome_sender(&self) -> Result<&LeafNode, LibraryError> {
-        self.group.welcome_sender()
+        let sender_index = self.welcome_sender_index();
+        self.public_group
+            .leaf(sender_index)
+            .ok_or(LibraryError::custom(
+                "no leaf with given welcome sender index exists",
+            ))
     }
 
     /// Consumes the [`StagedWelcome`] and returns the respective [`MlsGroup`].
@@ -362,24 +521,35 @@ impl StagedWelcome {
         self,
         provider: &Provider,
     ) -> Result<MlsGroup, WelcomeError<Provider::StorageError>> {
-        let mut group = self.group.into_core_group(provider)?;
-        group.set_max_past_epochs(self.mls_group_config.max_past_epochs);
+        // If we got a path secret, derive the path (which also checks if the
+        // public keys match) and store the derived keys in the key store.
+        let group_keypairs = if let Some(path_keypairs) = self.path_keypairs {
+            let mut keypairs = vec![self.key_package_bundle.encryption_key_pair()];
+            keypairs.extend_from_slice(&path_keypairs);
+            keypairs
+        } else {
+            vec![self.key_package_bundle.encryption_key_pair()]
+        };
 
-        let mls_group = MlsGroup {
+        let mut mls_group = MlsGroup {
             mls_group_config: self.mls_group_config,
-            group,
             own_leaf_nodes: vec![],
             aad: vec![],
             group_state: MlsGroupState::Operational,
+            public_group: self.public_group,
+            group_epoch_secrets: self.group_epoch_secrets,
+            own_leaf_index: self.own_leaf_index,
+            message_secrets_store: self.message_secrets_store,
+            resumption_psk_store: self.resumption_psk_store,
         };
 
-        provider
-            .storage()
-            .write_mls_join_config(mls_group.group_id(), &mls_group.mls_group_config)
+        mls_group
+            .store_epoch_keypairs(provider.storage(), group_keypairs.as_slice())
             .map_err(WelcomeError::StorageError)?;
-        provider
-            .storage()
-            .write_group_state(mls_group.group_id(), &MlsGroupState::Operational)
+        mls_group.set_max_past_epochs(mls_group.mls_group_config.max_past_epochs);
+
+        mls_group
+            .store(provider.storage())
             .map_err(WelcomeError::StorageError)?;
 
         Ok(mls_group)
