+++ openmls/src/group/mls_group/errors.rs
@@ -45,6 +45,9 @@ pub enum NewGroupError<StorageError> {
     /// Invalid extensions set in configuration
     #[error("Invalid extensions set in configuration")]
     InvalidExtensions(#[from] InvalidExtensionError),
+    /// See [`PskError`] for more details.
+    #[error(transparent)]
+    Psk(#[from] PskError),
 }
 
 /// EmptyInput error
