+++ openmls/src/group/mls_group/exporting.rs
@@ -1,6 +1,11 @@
-use openmls_traits::signatures::Signer;
+use errors::{ExportGroupInfoError, ExportSecretError};
+use openmls_traits::{crypto::OpenMlsCrypto, signatures::Signer};
 
-use crate::{group::errors::ExporterError, schedule::EpochAuthenticator, storage::OpenMlsProvider};
+use crate::{
+    ciphersuite::HpkePublicKey,
+    schedule::{EpochAuthenticator, ResumptionPskSecret},
+    storage::OpenMlsProvider,
+};
 
 use super::*;
 
@@ -21,14 +26,17 @@ impl MlsGroup {
     ) -> Result<Vec<u8>, ExportSecretError> {
         let crypto = provider.crypto();
 
+        if key_length > u16::MAX.into() {
+            log::error!("Got a key that is larger than u16::MAX");
+            return Err(ExportSecretError::KeyLengthTooLong);
+        }
+
         if self.is_active() {
             Ok(self
-                .group
-                .export_secret(crypto, label, context, key_length)
-                .map_err(|e| match e {
-                    ExporterError::LibraryError(e) => e.into(),
-                    ExporterError::KeyLengthTooLong => ExportSecretError::KeyLengthTooLong,
-                })?)
+                .group_epoch_secrets
+                .exporter_secret()
+                .derive_exported_secret(self.ciphersuite(), crypto, label, context, key_length)
+                .map_err(LibraryError::unexpected_crypto_error)?)
         } else {
             Err(ExportSecretError::GroupStateError(
                 MlsGroupStateError::UseAfterEviction,
@@ -54,7 +62,7 @@ impl MlsGroup {
     /// Returns a resumption psk for a given epoch. If no resumption psk
     /// is available for that epoch,  `None` is returned.
     pub fn get_past_resumption_psk(&self, epoch: GroupEpoch) -> Option<&ResumptionPskSecret> {
-        self.group.resumption_psk_store.get(epoch)
+        self.resumption_psk_store.get(epoch)
     }
 
     /// Export a group info object for this group.
@@ -64,9 +72,56 @@ impl MlsGroup {
         signer: &impl Signer,
         with_ratchet_tree: bool,
     ) -> Result<MlsMessageOut, ExportGroupInfoError> {
-        Ok(self
-            .group
-            .export_group_info(provider.crypto(), signer, with_ratchet_tree)?
-            .into())
+        let extensions = {
+            let ratchet_tree_extension = || {
+                Extension::RatchetTree(RatchetTreeExtension::new(
+                    self.public_group().export_ratchet_tree(),
+                ))
+            };
+
+            let external_pub_extension = || -> Result<Extension, ExportGroupInfoError> {
+                let external_pub = self
+                    .group_epoch_secrets()
+                    .external_secret()
+                    .derive_external_keypair(provider.crypto(), self.ciphersuite())
+                    .map_err(LibraryError::unexpected_crypto_error)?
+                    .public;
+                Ok(Extension::ExternalPub(ExternalPubExtension::new(
+                    HpkePublicKey::from(external_pub),
+                )))
+            };
+
+            if with_ratchet_tree {
+                Extensions::from_vec(vec![ratchet_tree_extension(), external_pub_extension()?])
+                    .map_err(|_| {
+                        LibraryError::custom(
+                            "There should not have been duplicate extensions here.",
+                        )
+                    })?
+            } else {
+                Extensions::single(external_pub_extension()?)
+            }
+        };
+
+        // Create to-be-signed group info.
+        let group_info_tbs = GroupInfoTBS::new(
+            self.context().clone(),
+            extensions,
+            self.message_secrets()
+                .confirmation_key()
+                .tag(
+                    provider.crypto(),
+                    self.ciphersuite(),
+                    self.context().confirmed_transcript_hash(),
+                )
+                .map_err(LibraryError::unexpected_crypto_error)?,
+            self.own_leaf_index(),
+        );
+
+        // Sign to-be-signed group info.
+        let group_info = group_info_tbs
+            .sign(signer)
+            .map_err(|_| LibraryError::custom("Signing failed"))?;
+        Ok(group_info.into())
     }
 }
