+++ openmls/src/group/mls_group/proposal.rs
@@ -1,10 +1,9 @@
 use openmls_traits::{signatures::Signer, storage::StorageProvider as _, types::Ciphersuite};
 
 use super::{
-    core_group::create_commit_params::CreateCommitParams,
-    errors::{ProposalError, ProposeAddMemberError, ProposeRemoveMemberError},
+    errors::{ProposalError, ProposeAddMemberError, ProposeRemoveMemberError, RemoveProposalError},
     CreateGroupContextExtProposalError, CustomProposal, GroupContextExtensionProposal, MlsGroup,
-    MlsGroupState, PendingCommitState, Proposal, RemoveProposalError,
+    MlsGroupState, PendingCommitState, Proposal,
 };
 use crate::{
     binary_tree::LeafNodeIndex,
@@ -12,7 +11,10 @@ use crate::{
     credentials::Credential,
     extensions::Extensions,
     framing::MlsMessageOut,
-    group::{errors::CreateAddProposalError, GroupId, QueuedProposal},
+    group::{
+        core_group::create_commit_params::CreateCommitParams, errors::CreateAddProposalError,
+        GroupId, QueuedProposal,
+    },
     key_packages::KeyPackage,
     messages::{group_info::GroupInfo, proposals::ProposalOrRefType},
     prelude::LibraryError,
@@ -72,9 +74,7 @@ macro_rules! impl_propose_fun {
         ) -> Result<(MlsMessageOut, ProposalRef), ProposalError<Provider::StorageError>> {
             self.is_operational()?;
 
-            let proposal = self
-                .group
-                .$group_fun(self.framing_parameters(), value, signer)?;
+            let proposal = self.$group_fun(self.framing_parameters(), value, signer)?;
 
             let queued_proposal = QueuedProposal::from_authenticated_content(
                 self.ciphersuite(),
@@ -87,7 +87,7 @@ macro_rules! impl_propose_fun {
             log::trace!("Storing proposal in queue {:?}", queued_proposal);
             provider
                 .storage()
-                .queue_proposal(self.group.group_id(), &proposal_ref, &queued_proposal)
+                .queue_proposal(self.group_id(), &proposal_ref, &queued_proposal)
                 .map_err(ProposalError::StorageError)?;
             self.proposal_store_mut().add(queued_proposal);
 
@@ -238,7 +238,6 @@ impl MlsGroup {
         self.is_operational()?;
 
         let add_proposal = self
-            .group
             .create_add_proposal(self.framing_parameters(), key_package.clone(), signer)
             .map_err(|e| match e {
                 CreateAddProposalError::LibraryError(e) => e.into(),
@@ -279,7 +278,6 @@ impl MlsGroup {
         self.is_operational()?;
 
         let remove_proposal = self
-            .group
             .create_remove_proposal(self.framing_parameters(), member, signer)
             .map_err(|_| ProposeRemoveMemberError::UnknownMember)?;
 
@@ -314,7 +312,6 @@ impl MlsGroup {
     {
         // Find the user for the credential first.
         let member_index = self
-            .group
             .public_group()
             .members()
             .find(|m| &m.credential == member)
@@ -339,7 +336,6 @@ impl MlsGroup {
     ) -> Result<(MlsMessageOut, ProposalRef), ProposalError<Provider::StorageError>> {
         // Find the user for the credential first.
         let member_index = self
-            .group
             .public_group()
             .members()
             .find(|m| &m.credential == member)
@@ -366,7 +362,7 @@ impl MlsGroup {
     ) -> Result<(MlsMessageOut, ProposalRef), ProposalError<Provider::StorageError>> {
         self.is_operational()?;
 
-        let proposal = self.group.create_group_context_ext_proposal::<Provider>(
+        let proposal = self.create_group_context_ext_proposal::<Provider>(
             self.framing_parameters(),
             extensions,
             signer,
@@ -420,7 +416,7 @@ impl MlsGroup {
             .framing_parameters(self.framing_parameters())
             .inline_proposals(inline_proposals)
             .build();
-        let create_commit_result = self.group.create_commit(params, provider, signer)?;
+        let create_commit_result = self.create_commit(params, provider, signer)?;
 
         let mls_messages = self.content_to_mls_message(create_commit_result.commit, provider)?;
 
@@ -440,7 +436,7 @@ impl MlsGroup {
             mls_messages,
             create_commit_result
                 .welcome_option
-                .map(|w| MlsMessageOut::from_welcome(w, self.group.version())),
+                .map(|w| MlsMessageOut::from_welcome(w, self.version())),
             create_commit_result.group_info,
         ))
     }
